requirements.txt
# Google Cloud Platform SDK
google-cloud-compute>=1.14.0
google-cloud-monitoring>=2.15.0
google-cloud-logging>=3.5.0
google-cloud-storage>=2.10.0
google-cloud-resource-manager>=1.10.0
google-auth>=2.23.0

# 環境変数管理
python-dotenv>=1.0.0

# Ansible
ansible>=8.0.0
ansible-core>=2.15.0

# HTTP クライアント
requests>=2.31.0

# CLI ツール
click>=8.1.0

# ロギング・モニタリング
structlog>=23.1.0

# 設定管理
pyyaml>=6.0.1

# テスト
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# コード品質
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# AIエージェントフレームワーク（オプション）
# langchain>=0.1.0
# anthropic>=0.7.0
# openai>=1.0.0

